update to 1.19.4

+++ src/main/java/kaptainwutax/seedcrackerX/finder/decorator/EndPillarsFinder.java
@@ -38,7 +38,7 @@ public class EndPillarsFinder extends Finder {
                 x = Math.round(x);
                 z = Math.round(z);
             }
-            this.bedrockMarkers[i] = new BedrockMarkerFinder(this.world, new ChunkPos(new BlockPos(x, 0, z)), new BlockPos(x, 0, z));
+            this.bedrockMarkers[i] = new BedrockMarkerFinder(this.world, new ChunkPos(new BlockPos((int)x, 0, (int)z)), new BlockPos((int)x, 0, (int)z));
         }
     }
